[persistence] Support for limit and offset filter on get all endpoints (#1024)
* change limit and offset data type to Long in DaoFilter

* unlock limit and offset on getAll

* Handle case where limit and offset are the only filters

* test cases

* added license to the test file

* removed check and added doc to filter

* fixes

* Add negative tests

* cosmetic fixes

* test rename

* simplifying tests

* Unit tests on persistence layer for pagination

* added comment

* test bug fix

+++ thirdeye-persistence/src/main/java/ai/startree/thirdeye/datalayer/bao/TaskManagerImpl.java
@@ -247,7 +247,7 @@ public class TaskManagerImpl implements TaskManager {
     final long startTime = System.nanoTime();
     final List<TaskDTO> tasksToBeDeleted = filter(new DaoFilter()
         .setPredicate(Predicate.LT("createTime", formattedDate))
-        .setLimit(limit)
+        .setLimit((long) limit)
     );
 
     /* Delete each task */
